+++ packages/lumx-react/src/utils/date/isSameDay.ts
@@ -1,0 +1,7 @@
+/**
+ * Check `date1` is on the same day as `date2`.
+ */
+export const isSameDay = (date1: Date, date2: Date) =>
+    date1.getFullYear() === date2.getFullYear() &&
+    date1.getMonth() === date2.getMonth() &&
+    date1.getDate() === date2.getDate();