+++ packages/lumx-react/src/utils/locale/getCurrentLocale.ts
@@ -1,0 +1,4 @@
+/**
+ * Get current browser locale.
+ */
+export const getCurrentLocale = (): string => navigator.languages?.[0] || navigator.language;